Clean up new webpack code
